Update User ID field type in migrations
Updated user_id feild type from 'integer' to 'unsignedBigInteger' in line with Laravel 5.8 updates.

+++ migrations/2018_04_04_134349_create_users_roles_table.php
@@ -14,7 +14,7 @@ class CreateUsersRolesTable extends Migration
     public function up()
     {
         Schema::create('users_roles', function (Blueprint $table) {
-            $table->integer('user_id')->unsigned();
+            $table->unsignedBigInteger('user_id')->unsigned();
             $table->integer('role_id')->unsigned();
 
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
